clean up genre list code
